+++ mathics/builtin/assignment.py
@@ -231,15 +231,15 @@ class _SetOperator(object):
             allowed_names = [focus.get_lookup_name()]
             if allow_custom_tag:
                 for leaf in focus.get_leaves():
-                    if not leaf.is_symbol() and leaf.get_head_name() in (
+                    if type(leaf) is not Symbol and leaf.get_head_name() in (
                         "System`HoldPattern",
                     ):
                         leaf = leaf.leaves[0]
-                    if not leaf.is_symbol() and leaf.get_head_name() in (
+                    if type(leaf) is not Symbol and leaf.get_head_name() in (
                         "System`Pattern",
                     ):
                         leaf = leaf.leaves[1]
-                    if not leaf.is_symbol() and leaf.get_head_name() in (
+                    if type(leaf) is not Symbol and leaf.get_head_name() in (
                         "System`Blank",
                         "System`BlankSequence",
                         "System`BlankNullSequence",