+++ mathics/builtin/optimization.py
@@ -76,7 +76,7 @@ class Minimize(Builtin):
         vars = vars.leaves
         for var in vars:
             if (
-                (var.is_atom() and not var.is_symbol())
+                (var.is_atom() and type(var) is not Symbol)
                 or head_name in ("System`Plus", "System`Times", "System`Power")  # noqa
                 or "System`Constant" in var.get_attributes(evaluation.definitions)
             ):
@@ -155,7 +155,7 @@ class Minimize(Builtin):
         vars = vars.leaves
         for var in vars:
             if (
-                (var.is_atom() and not var.is_symbol())
+                (var.is_atom() and type(var) is not Symbol)
                 or head_name in ("System`Plus", "System`Times", "System`Power")  # noqa
                 or "System`Constant" in var.get_attributes(evaluation.definitions)
             ):